• Adds home.js into a foler in pages • Moves home.js style into home.css created from app.css
