+++ anime-archive/src/pages/Home/Home.css
@@ -1,0 +1,6 @@
+header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 10px;
+}
