+++ anime-archive/src/pages/Home/Home.js
@@ -1,9 +1,9 @@
-import Logo from "../components/logo/Logo";
-import Carousel from "../components/carousel/Carousel.js";
-import Section from "../components/section/Section.js";
-import { sectionData } from "../utils/sectionData";
-import SearchImg from "../assets/icons/searchIcon.png";
-import "../App.css";
+import Logo from "../../components/logo/Logo";
+import Carousel from "../../components/carousel/Carousel.js";
+import Section from "../../components/section/Section.js";
+import { sectionData } from "../../utils/sectionData";
+import SearchImg from "../../assets/icons/searchIcon.png";
+import "./Home.css";
 
 export default function Home() {
   return (
